取web数据地址已迁移至数据库

+++ src/main/java/com/newsmetro/service/UserService.java
@@ -77,7 +77,9 @@ public class UserService {
 //		for(int i=0;i<list.size();i++){
 //			array[i]=list.get(i);
 //		}
-		userMapper.deleteByIdList(list);
+        if(list!=null&&list.size()>0){
+            userMapper.deleteByIdList(list);
+        }
 	}
 	
 	public void update(User user){
